add prop validations.
