+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\Microverse\\physician-appointments\\src\\index.js":"1","C:\\Microverse\\physician-appointments\\src\\reportWebVitals.js":"2","C:\\Microverse\\physician-appointments\\src\\components\\App.js":"3","C:\\Microverse\\physician-appointments\\src\\components\\Login.js":"4"},{"size":512,"mtime":1612264066053,"results":"5","hashOfConfig":"6"},{"size":373,"mtime":1612264066053,"results":"7","hashOfConfig":"6"},{"size":1213,"mtime":1612298095821,"results":"8","hashOfConfig":"6"},{"size":2035,"mtime":1612297077326,"results":"9","hashOfConfig":"6"},{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"p1lrsy",{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","errorCount":0,"warningCount":3,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"16","messages":"17","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\Microverse\\physician-appointments\\src\\index.js",[],"C:\\Microverse\\physician-appointments\\src\\reportWebVitals.js",[],"C:\\Microverse\\physician-appointments\\src\\components\\App.js",["18","19","20"],"C:\\Microverse\\physician-appointments\\src\\components\\Login.js",["21","22"],{"ruleId":"23","severity":1,"message":"24","line":14,"column":3,"nodeType":"25","messageId":"26","endLine":14,"endColumn":14},{"ruleId":"23","severity":1,"message":"24","line":32,"column":7,"nodeType":"25","messageId":"26","endLine":32,"endColumn":18},{"ruleId":"23","severity":1,"message":"24","line":34,"column":7,"nodeType":"25","messageId":"26","endLine":34,"endColumn":18},{"ruleId":"23","severity":1,"message":"24","line":36,"column":7,"nodeType":"25","messageId":"26","endLine":36,"endColumn":18},{"ruleId":"23","severity":1,"message":"24","line":45,"column":7,"nodeType":"25","messageId":"26","endLine":45,"endColumn":18},"no-console","Unexpected console statement.","MemberExpression","unexpected"]
+[{"C:\\Microverse\\physician-appointments\\src\\index.js":"1","C:\\Microverse\\physician-appointments\\src\\reportWebVitals.js":"2","C:\\Microverse\\physician-appointments\\src\\components\\App.js":"3","C:\\Microverse\\physician-appointments\\src\\components\\Login.js":"4","C:\\Microverse\\physician-appointments\\src\\components\\Appointments.js":"5"},{"size":556,"mtime":1612299945585,"results":"6","hashOfConfig":"7"},{"size":373,"mtime":1612264066053,"results":"8","hashOfConfig":"7"},{"size":857,"mtime":1612301418895,"results":"9","hashOfConfig":"7"},{"size":2357,"mtime":1612302910250,"results":"10","hashOfConfig":"7"},{"size":861,"mtime":1612303689629,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1azknno",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"18","messages":"19","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Microverse\\physician-appointments\\src\\index.js",[],"C:\\Microverse\\physician-appointments\\src\\reportWebVitals.js",[],"C:\\Microverse\\physician-appointments\\src\\components\\App.js",["22","23"],"C:\\Microverse\\physician-appointments\\src\\components\\Login.js",["24","25"],"C:\\Microverse\\physician-appointments\\src\\components\\Appointments.js",[],{"ruleId":"26","severity":1,"message":"27","line":14,"column":7,"nodeType":"28","messageId":"29","endLine":14,"endColumn":18},{"ruleId":"26","severity":1,"message":"27","line":16,"column":7,"nodeType":"28","messageId":"29","endLine":16,"endColumn":18},{"ruleId":"26","severity":1,"message":"27","line":39,"column":7,"nodeType":"28","messageId":"29","endLine":39,"endColumn":18},{"ruleId":"26","severity":1,"message":"27","line":48,"column":7,"nodeType":"28","messageId":"29","endLine":48,"endColumn":18},"no-console","Unexpected console statement.","MemberExpression","unexpected"]
